внесла правки в дз
