+++ js/form.js
@@ -1,9 +1,15 @@
 import { isEscapeKey } from './util.js';
-import { resetValidation } from './validation.js';
+import { resetValidation, validateForm } from './validation.js';
 import { resetScale } from './size-picture.js';
 import { resetEffect } from './picture-filter.js';
+import { sendData } from './api.js';
 
 const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
+
+const SubmitButtonText = {
+  IDLE: 'ОПУБЛИКОВАТЬ',
+  SENDING: 'Отправляю...'
+};
 
 const imgUpload = document.getElementById('upload-file');
 const uploadCancel = document.getElementById('upload-cancel');
@@ -11,6 +17,28 @@ const imgUploadOverlay = document.querySelector('.img-upload__overlay');
 const imgUploadPreview = document.querySelector('.img-upload__preview img');
 const imgUploadInput = document.querySelector('.img-upload__start input[type=file]');
 const effectsPreview = document.querySelectorAll('.effects__preview');
+const submitButton = document.querySelector('.img-upload__submit');
+const imgForm = document.querySelector('.img-upload__form');
+const textHashtags = document.querySelector('.text__hashtags');
+const textDescription = document.querySelector('.text__description');
+
+const isFocusid = () =>
+  document.activeElement === textHashtags ||
+  document.activeElement === textDescription;
+
+textHashtags.addEventListener('keydown', (evt) => {
+
+  if (evt.key === 'Escape'){
+    evt.stopPropagation();
+  }
+});
+
+textDescription.addEventListener('keydown', (evt) => {
+
+  if (evt.key === 'Escape'){
+    evt.stopPropagation();
+  }
+});
 
 const onDocumentKeydown = (evt) => {
 
@@ -59,5 +87,39 @@ const setPhotoListener = () => {
   });
 };
 
-export {setPhotoListener, openUploadModal, closeUploadModal};
+const blockSubmitButton = () => {
+  submitButton.disabled = true;
+  submitButton.textContent = SubmitButtonText.SENDING;
+};
+
+const unblockSubmitButton = () => {
+  submitButton.disabled = false;
+  submitButton.textContent = SubmitButtonText.IDLE;
+};
+
+const onSendSuccess = () => {
+  closeUploadModal();
+};
+
+const onSendFail = () => {
+  console.log('ошибка');
+};
+
+const setUserFormSubmit = () => {
+
+  imgForm.addEventListener('submit', (evt) => {
+    evt.preventDefault();
+
+    const isValid = validateForm();
+    if (isValid) {
+      blockSubmitButton();
+      sendData(new FormData(evt.target))
+        .then(onSendSuccess)
+        .catch(onSendFail)
+        .finally(unblockSubmitButton);
+    }
+  });
+};
+
+export {setPhotoListener, setUserFormSubmit};
