+++ js/main.js
@@ -1,11 +1,12 @@
-import {setupValidation, setUserFormSubmit} from './validation.js';
-import {setPhotoListener, closeUploadModal} from './form.js';
+import {setupValidation} from './validation.js';
+import {setPhotoListener, setUserFormSubmit} from './form.js';
 import { initScale } from './size-picture.js';
 import { initSlider } from './picture-filter.js';
 import { renderPictures } from './render-pictures.js';
 import { getData} from './api.js';
 import { showAlert } from './util.js';
 import {renderGallery} from './render-gallery.js';
+import { showSuccessMessage } from './show-modal-message.js';
 
 const PICTURE_COUNT = 25;
 
@@ -13,6 +14,8 @@ setPhotoListener();
 setupValidation();
 initScale();
 initSlider();
+setUserFormSubmit();
+// showSuccessMessage();
 
 getData()
   .then((picture) => {
@@ -24,5 +27,3 @@ getData()
       showAlert(err.message);
     }
   );
-
-setUserFormSubmit(closeUploadModal);
