--- js/sending-message.js
@@ -1,45 +1,0 @@
-const success = document.getElementById('success').content;
-const newItemTemplateSusses = success.querySelector('.success');
-const imgUploadOverlay = document.querySelector('.img-upload__overlay');
-const successButton = success.querySelector('.success__button');
-const error = document.getElementById('error').content;
-const newTemplateError = error.querySelector('.error');
-const textHashtags = document.querySelector('.text__hashtags');
-const textDescription = document.querySelector('.text__description');
-
-
-const isFocusid = () =>
-  document.activeElement === textHashtags ||
-  document.activeElement === textDescription;
-
-textHashtags.addEventListener('keydown', (evt) => {
-
-  if (evt.key === 'Escape'){
-    evt.stopPropagation();
-  }
-});
-
-textDescription.addEventListener('keydown', (evt) => {
-
-  if (evt.key === 'Escape'){
-    evt.stopPropagation();
-  }
-});
-
-
-function shippedMessageSuccess () {
-  document.body.classList.remove('modal-open');
-  imgUploadOverlay.classList.add('hidden');
-  const cloneTemplateSuccess = newItemTemplateSusses.cloneNode(true);
-  document.body.appendChild(cloneTemplateSuccess);
-  successButton.addEventListener('click', () => {
-    cloneTemplateSuccess.remove();
-  });
-}
-
-function shippedMessageError () {
-  const cloneTemplateError = newTemplateError.cloneNode(true);
-  document.body.appendChild(cloneTemplateError);
-}
-
-export {shippedMessageSuccess, shippedMessageError};
